Its sooooo groossssss.
But it kinda works. It now can sort of cast spells. Not based on input yet.
And player health is super broken. Yipee

+++ Task4/Task4/Task4/String.h
@@ -30,8 +30,8 @@ public:
 	String ToLower();
 	String ToUpper();
 
-	size_t Find(const String findString);
-	size_t Find(const int startIndex, const String findString);
+	int Find(const String findString);
+	int Find(const int startIndex, const String findString);
 
 	String& Replace(const String findString, const String replaceString);
